+++ prover/src/tests/mod.rs
@@ -11,10 +11,8 @@ pub mod constraints64_tests;
 #[cfg(test)]
 pub mod cpu_tests;
 #[cfg(test)]
-pub mod cpu_vm_tests;
-#[cfg(test)]
 pub mod lt_bus_tests;
 #[cfg(test)]
 pub mod lt_tests;
 #[cfg(test)]
-pub mod vm_prover_tests;
+pub mod prove_elfs_tests;
